Actually fixed cyclic dependency this time.
